// src/lib/interfaces/maps/ponds.ts
export interface PondCoordinate {
  lng: number;
  lat: number;
}


export interface StarterBox {
  id: number;
  title: string;
  alias_starter_title: string | null;
  status: string
  mac_address: string;
  gateways?: Gateway
}

export interface UserDevices {
  id: number;
  title: string;
  alias_starter_title: string | null;
  connected_nodes: string[];
  capable_motors: number;
}


export interface Faults {
  fault_code: number;
  fault_description: string;
  time_stamp: string;
  power_present: string;
}

export interface StarterBoxParameters {
  id: number;
  starter_id: number;
  motor_id: number;
  current_i1: number;
  current_i2: number;
  current_i3: number;
  line_voltage_vry: number;
  line_voltage_vyb: number;
  line_voltage_vbr: number;
  power_present?: string;
  fault_code?: number;
  fault_description?: string;
}

export interface Motor {
  id: number;
  title: string;
  state: number;
  starter_id: number | null;
  motor_ref_id: string | null;
  status: string
  mode: string
  hp: string;
  motor_coordinates: {
    lat: number;
    lng: number;
  };
  starterBoxParameters: StarterBoxParameters[];
  starterBox: StarterBox;
  faults: Faults;
}

export interface Pond {
  id: number;
  title: string;
  location_id: number;
  pond_order: number;
  pond_coordinates: PondCoordinate[];
  acres: string;
  status: "ACTIVE" | "INACTIVE";
  motors: Motor[];
  gateway: Gateway;
  location: Location;
  total_motors_count: number;
  motors_on_count: number;
  motors_off_count: number;
  motors_online_count: number;
  motors_offline_count: number;
  power_on_count: number;
  power_off_count: number;
  auto_count: number;
  manual_count: number;
  alert_count: number;
}
// ...
export interface Location {
  id: number;
  name: string;
  title?: string;
  location_coordinates?: {
    lat: number;
    lng: number;
  }
  gateway_count?: number
}